IntialCommit
The plan for this repository will be that it will have 2 parts
Server : which serves as a tremplates for all SRBNP Based Servers to be made of.
and
Client: which serves as a tremplates for all SRBNP Based Servers to be made of.
This project is inspired by Spectrom_Server's Spectrom Network Protocol, they however will not use the same network protocol, for now at least.
Currently i only added a thread poll (thpool) and a simple TCP server that doesnt accept Incoming connections, Epoll is in the making.
i also made a make file that deploys a service directly, only works if you are using a system V init linux services manager for now