Adeed HashTables, preparing to remove SQLite
Added a hashtable that will be used to save the clients, included Math library. the hashtable algorithm will use jenkins's locate3.c [See More](https://www.burtleburtle.net/bob/c/lookup3.c) as hashing algorithm and Linked list for collision handling.
Preparing to remove SQlite and be replaced by the hashtable for better performance, and since it's is planned to be a library it wouldn't make sence to include a whole SQL Library into it. and make as lightweight as possible.