Client/Java Simple Connector ( Unvifinished )
The current Java part of the Client part will serve as a Connector, there are plans explaining how to use this Connector  and how to make projects with it.
it has Two parts as of now
[ SRBNP Package  📦 ]
- Here is lies all principal functioning of the connector
- All class difinitions
[ SRBNP.Exceptions Package 📦 ]
- Saved all Custom Exceptions

Note: the structuring will change heavily during the inital developpement.
The current connector is Only capable of Sending "Requests" one at a time and then put them in a "Unresponded Requests" List, which would be then removed from that list once the server responds to a certain Request. However the latter part where it receives responses is unavailable as of now, and this connector.