Server/C is now a Library!
after hours of continous work, SRBNP is now fully a library though it lacks the capability to be installed directly now, and is a Static library.

the make file has two principal rules
 make clean : removes any object/header/library files
 make build : cleans and then compiles the library
 make test : compiles a simple program to be used for testing purposes.

all Code stored in Server/C/SRBNP/ has been moved to Server/C/src
and _Import.h changed to SRBNP.h
the library remains unfinished and not recommended for use.

Upon build aa library will be saved in Server/C/lib/ as " libsrbnp.a " Check 'test' rule in makefile for proper explanation on how to compile with it for now.

changes made in all headerfiles to accommodate  the change to a static library.