Server/C Changing OPCODE structre
In src/opcode  📂:
changing the structre of opcodes to a more dynamicly declared opcodes,
an OPCODE now consists of 4 parts each 1byte  (Keeps same size as before the change ) which are:
1 -  Category
2 - Sub-Category
3 - Command
4 - Detail

Each opcode command must be defined inside  a sub category,  the latter must be defined inside a category

a "Detail" is used to define a reason or justification for the command.
by example
In Base Category ( 0x00 )
In Base SubCategory ( 0x00 )
In Invalid Command ( 0xFF )

there  are 3 possible "Detail"s ( or simply justifications):

Note: "Detail"s with value 0x00 MUST mean that the Command was sent without any justification ( when command sends a generic command response  or no "Detail" needed for said Command).

1 .._UNDETAILED ( 0x00 ) which mean there is an undetailed reason for the invalid request.
2 .._UNREGISTRED_OPCODE ( 0x01 ) means the request was invalidated due to an invalid opcode
3 .._WRONG_PROTOCOL ( 0x02 ) means the client had sent a valid opcode in the wrong place. ( Maybe due version change )

PS: This source will be documented in the upcoming commits.