Properly Set licenses & Fixed Typo
The License (LGPL v2.1 ) is actually mean Just for Server/C and Client/Java and README.md, so i set it so that each of them had the license(not the whole Project) as it should've been from the start.

Now Server/C ( Or the C Version of the SNF Server library "libsnf" )
Client/Java ( Or the Java Version of the SNF Client Library )
are both under LGPL v2.1 as of this Commit, But see the license on their folder to know what license they Currently have

In Server/C :
Fixed a Typod where it should've just received 36 bytes of the Client's GUID, but i left it at 37 byte accidentally.
IN Client/Java:
As per the change of Server/C this client is depricated, as planned, in the future i will temporarly stop this version until further notice. a new Client Library will come in the upcoming commits