Fixed Typos, and Improuved Variable System
Server/C:

Fixed typos and errors in thpool.test

Improuved variable system by not allowing variables to be changed if certain conditions are not met ( for now just checks in case the Value contains an invalid value depending on the variable) and ability to restrict the modification of some variables manually.
Added 'vars'  checks to test the recent improuvements.

Also started to improuve the error handling system