Go/Client Untested Client implementation.
Tests shall come after a few days of university tests. Good progress

also a minor  change in protocol and instead of server info in Initial Server Request, being the first "Special Metadata"  I made it as an argument to know how long is the server information due to how versatile it is.